feat: split swap deployment script
